+++ db/migrate/20251124090100_add_scoring_system_id_to_game_events.rb
@@ -1,0 +1,7 @@
+class AddScoringSystemIdToGameEvents < ActiveRecord::Migration[7.1]
+  def change
+    add_reference :game_events, :scoring_system, foreign_key: { to_table: :game_scoring_systems }
+  end
+end
+
+