+++ db/migrate/20251124090200_add_scoring_system_id_to_tournaments.rb
@@ -1,0 +1,7 @@
+class AddScoringSystemIdToTournaments < ActiveRecord::Migration[7.1]
+  def change
+    add_reference :tournaments, :scoring_system, foreign_key: { to_table: :game_scoring_systems }
+  end
+end
+
+